Move callbacks into own submodule
