+++ skorch/__init__.py
@@ -8,7 +8,6 @@ from .net import NeuralNetClassifier
 from .net import NeuralNetRegressor
 
 from . import callbacks
-from . import lr_scheduler
 
 
 __all__ = [
@@ -16,7 +15,7 @@ __all__ = [
     'NeuralNet',
     'NeuralNetClassifier',
     'NeuralNetRegressor',
-    'callbacks'
+    'callbacks',
 ]
